refactor: rename database.ts to databaseSetup.ts
